pyproject.toml
[tool.poetry]
name = "chat"
version = "0.1.0"
description = ""
authors = ["Ben Southcott <benjamin.southcott@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
websockets = "^12.0"
aio-pika = "^9.4.1"
fastapi = "^0.111.0"
pyjwt = "^2.8.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
motor = "^3.4.0"
redis = "^5.0.4"
cryptography = "^42.0.7"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
ipython = "^8.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
